requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
sqlalchemy==2.0.19
mysqlclient==2.1.1
flake8==6.0.0
bandit==1.7.5
pylint==2.17.4
pytest==7.4.0
black==23.11.0